+++ src/components/PurchaseOrder/OngoingOgderInfo/index.js
@@ -1,0 +1,2 @@
+export { default as OngoingOrderInfoView } from './OngoingOrderInfoView';
+export { default as OngoingInfoForm } from './OngoingInfoForm';