+++ src/components/PurchaseOrder/renewals/index.js
@@ -1,4 +1,0 @@
-import RenewalsView from './RenewalsView';
-import RenewalsForm from './RenewalsForm';
-
-export { RenewalsView, RenewalsForm };