+++ src/settings/OrderTemplates/OrderTemplateView/OrderTemplateView.js
@@ -21,7 +21,7 @@ import { FundDistributionView } from '@folio/stripes-acq-components';
 
 import { PODetailsView } from '../../../components/PurchaseOrder/PODetails';
 import { SummaryView } from '../../../components/PurchaseOrder/Summary';
-import { RenewalsView } from '../../../components/PurchaseOrder/renewals';
+import { OngoingOrderInfoView } from '../../../components/PurchaseOrder/OngoingOgderInfo';
 import { ItemView } from '../../../components/POLine/Item';
 import { POLineDetails } from '../../../components/POLine/POLineDetails';
 import { CostView } from '../../../components/POLine/Cost';
@@ -67,7 +67,7 @@ class OrderTemplateView extends Component {
       [ORDER_TEMPLATES_ACCORDION.PO_INFO]: false,
       [ORDER_TEMPLATES_ACCORDION.PO_TAGS]: false,
       [ORDER_TEMPLATES_ACCORDION.PO_SUMMARY]: false,
-      [ORDER_TEMPLATES_ACCORDION.PO_RENEWALS]: false,
+      [ORDER_TEMPLATES_ACCORDION.PO_ONGOING]: false,
       [ORDER_TEMPLATES_ACCORDION.POL_ITEM_DETAILS]: false,
       [ORDER_TEMPLATES_ACCORDION.POL_DETAILS]: false,
       [ORDER_TEMPLATES_ACCORDION.POL_COST_DETAILS]: false,
@@ -228,10 +228,10 @@ class OrderTemplateView extends Component {
 
                 {isOngoing(orderType) && (
                   <Accordion
-                    label={ORDER_TEMPLATES_ACCORDION_TITLES[ORDER_TEMPLATES_ACCORDION.PO_RENEWALS]}
-                    id={ORDER_TEMPLATES_ACCORDION.PO_RENEWALS}
+                    label={ORDER_TEMPLATES_ACCORDION_TITLES[ORDER_TEMPLATES_ACCORDION.PO_ONGOING]}
+                    id={ORDER_TEMPLATES_ACCORDION.PO_ONGOING}
                   >
-                    <RenewalsView order={orderTemplate} />
+                    <OngoingOrderInfoView order={orderTemplate} />
                   </Accordion>
                 )}