+++ test/bigtest/interactors/order-details-page.js
@@ -55,7 +55,7 @@ export default interactor(class OrderDetailsPage {
   approveOrderButton = new Button('[data-test-approve-order-button]');
   closeOrderButton = new Button('[data-test-close-order-button]');
   reopenOrderButton = new Button('[data-test-reopen-order-button]');
-  renewalsAccordion = isPresent('#renewals');
+  renewalsAccordion = isPresent('#ongoing');
   billTo = new BillTo();
   shipTo = new ShipTo();
   workflowStatus = new WorkflowStatus();